--- a/slasher/rpc/BUILD.bazel
+++ b/slasher/rpc/BUILD.bazel
@@ -17,7 +17,6 @@ go_library(
         "@com_github_gogo_protobuf//types:go_default_library",
         "@com_github_pkg_errors//:go_default_library",
         "@com_github_prysmaticlabs_go_ssz//:go_default_library",
-        "@com_github_prysmaticlabs_prysm//proto/eth/v1alpha1:go_default_library",
         "@org_golang_google_grpc//codes:go_default_library",
         "@org_golang_google_grpc//status:go_default_library",
     ],