--- a/slasher/rpc/detect_update_min_max_span.go
+++ b/slasher/rpc/detect_update_min_max_span.go
@@ -4,7 +4,7 @@ import (
 	"context"
 	"fmt"
 
-	ethpb "github.com/prysmaticlabs/prysm/proto/eth/v1alpha1"
+	ethpb "github.com/prysmaticlabs/prysm/proto/beacon/rpc/v1"
 	"github.com/prysmaticlabs/prysm/shared/params"
 )