--- a/slasher/rpc/server.go
+++ b/slasher/rpc/server.go
@@ -2,6 +2,8 @@ package rpc
 
 import (
 	"context"
+	"fmt"
+	"sync"
 
 	"github.com/gogo/protobuf/proto"
 	"github.com/gogo/protobuf/types"
@@ -35,29 +37,47 @@ func (ss *Server) IsSlashableAttestation(ctx context.Context, req *ethpb.Indexed
 		return nil, err
 	}
 	atsSlashinngRes := &ethpb.AttesterSlashingResponse{}
+	at := make(chan []*ethpb.AttesterSlashing)
+	er := make(chan error)
+	var wg sync.WaitGroup
+	lastIdx := int64(-1)
 	for _, idx := range indices {
-		atts, err := ss.SlasherDB.DoubleVotes(tEpoch, idx, root[:], req)
-		if err != nil {
-			return nil, err
+		if int64(idx) <= lastIdx {
+			return nil, fmt.Errorf("indexed attestation contains repeated or non sorted ids")
 		}
-		if atts != nil && len(atts) > 0 {
-			atsSlashinngRes.AttesterSlashing = append(atsSlashinngRes.AttesterSlashing, atts...)
-		}
+		wg.Add(1)
+		go func(idx uint64) {
+			atts, err := ss.SlasherDB.DoubleVotes(tEpoch, idx, root[:], req)
+			if err != nil {
+				er <- err
+				wg.Done()
+				return
+			}
+			if atts != nil && len(atts) > 0 {
+				at <- atts
+			}
+			atts, err = ss.DetectSurroundVotes(ctx, idx, req)
+			if err != nil {
+				er <- err
+				wg.Done()
+				return
+			}
+			if atts != nil && len(atts) > 0 {
+				at <- atts
+			}
+			wg.Done()
+		}(idx)
 	}
-
-	for _, idx := range indices {
-		atts, err := ss.DetectSurroundVotes(ctx, req.Data.Source.Epoch, req.Data.Target.Epoch, idx)
-		if err != nil {
-			return nil, err
-		}
-		for _, ia := range atts {
-			atsSlashinngRes.AttesterSlashing = append(atsSlashinngRes.AttesterSlashing, &ethpb.AttesterSlashing{
-				Attestation_1: req,
-				Attestation_2: ia,
-			})
-		}
+	wg.Wait()
+	close(er)
+	close(at)
+	for e := range er {
+		err = fmt.Errorf(err.Error() + " : " + e.Error())
 	}
-	return atsSlashinngRes, nil
+	for atts := range at {
+		atsSlashinngRes.AttesterSlashing = append(atsSlashinngRes.AttesterSlashing, atts...)
+	}
+	return atsSlashinngRes, err
 }
 
 // IsSlashableBlock returns a proposer slashing if the block header submitted is
@@ -101,24 +121,34 @@ func (ss *Server) SlashableAttestations(req *types.Empty, server ethpb.Slasher_S
 
 // DetectSurroundVotes is a method used to return the attestation that were detected
 // by min max surround detection method.
-func (ss *Server) DetectSurroundVotes(ctx context.Context, source uint64, target uint64, validatorIdx uint64) ([]*ethpb.IndexedAttestation, error) {
-	minTargetEpoch, err := ss.DetectAndUpdateMinEpochSpan(ctx, source, target, validatorIdx)
+func (ss *Server) DetectSurroundVotes(ctx context.Context, validatorIdx uint64, req *ethpb.IndexedAttestation) ([]*ethpb.AttesterSlashing, error) {
+	spanMap, err := ss.SlasherDB.ValidatorSpansMap(validatorIdx)
 	if err != nil {
 		return nil, err
 	}
-	maxTargetEpoch, err := ss.DetectAndUpdateMaxEpochSpan(ctx, source, target, validatorIdx)
+	minTargetEpoch, spanMap, err := ss.DetectAndUpdateMinEpochSpan(ctx, req.Data.Source.Epoch, req.Data.Target.Epoch, validatorIdx, spanMap)
 	if err != nil {
 		return nil, err
 	}
-	var idxAtts []*ethpb.IndexedAttestation
+	maxTargetEpoch, spanMap, err := ss.DetectAndUpdateMaxEpochSpan(ctx, req.Data.Source.Epoch, req.Data.Target.Epoch, validatorIdx, spanMap)
+	if err != nil {
+		return nil, err
+	}
+	if err := ss.SlasherDB.SaveValidatorSpansMap(validatorIdx, spanMap); err != nil {
+		return nil, err
+	}
+	var as []*ethpb.AttesterSlashing
 	if minTargetEpoch > 0 {
 		attestations, err := ss.SlasherDB.IndexedAttestation(minTargetEpoch, validatorIdx)
 		if err != nil {
 			return nil, err
 		}
 		for _, ia := range attestations {
-			if ia.Data.Source.Epoch > source && ia.Data.Target.Epoch < target {
-				idxAtts = append(idxAtts, ia)
+			if ia.Data.Source.Epoch > req.Data.Source.Epoch && ia.Data.Target.Epoch < req.Data.Target.Epoch {
+				as = append(as, &ethpb.AttesterSlashing{
+					Attestation_1: req,
+					Attestation_2: ia,
+				})
 			}
 		}
 	}
@@ -128,10 +158,13 @@ func (ss *Server) DetectSurroundVotes(ctx context.Context, source uint64, target
 			return nil, err
 		}
 		for _, ia := range attestations {
-			if ia.Data.Source.Epoch < source && ia.Data.Target.Epoch > target {
-				idxAtts = append(idxAtts, ia)
+			if ia.Data.Source.Epoch < req.Data.Source.Epoch && ia.Data.Target.Epoch > req.Data.Target.Epoch {
+				as = append(as, &ethpb.AttesterSlashing{
+					Attestation_1: req,
+					Attestation_2: ia,
+				})
 			}
 		}
 	}
-	return idxAtts, nil
+	return as, nil
 }