[Part of] #3133

---

# Description

while trying to asses the db footprint of slasher attestations on local db it seemed i cant load it in reasonable time with attestation as it is not optimized enough.

- added concurrent calls 
- single put for single detection
- implemented transaction batching
Without batch(2 iterations):

<img width="1185" alt="Screen Shot 2019-12-04 at 10 40 52 AM" src="https://user-images.githubusercontent.com/5848741/70114483-9b94c800-1682-11ea-81e9-032e07cef06d.png">

With Batch(2 iterations):

<img width="1238" alt="Screen Shot 2019-12-04 at 10 42 49 AM" src="https://user-images.githubusercontent.com/5848741/70114544-d860bf00-1682-11ea-9e44-f8e7dd073857.png">

With batch (100 iterations):

<img width="1194" alt="Screen Shot 2019-12-04 at 10 44 52 AM" src="https://user-images.githubusercontent.com/5848741/70114643-22e23b80-1683-11ea-878a-aee58b6c9795.png">
Benchmark Before:

BenchmarkMinSpan/MinSpan_diff_2-4         	     100	  24873579 ns/op	   13967 B/op	     113 allocs/op
BenchmarkMinSpan/MinSpan_diff_10-4        	     100	  21771577 ns/op	   18562 B/op	     165 allocs/op
BenchmarkMinSpan/MinSpan_diff_100-4       	     552	  19274973 ns/op	   53920 B/op	     353 allocs/op
BenchmarkMinSpan/MinSpan_diff_1000-4      	    8571	  24328417 ns/op	  748390 B/op	    4483 allocs/op
BenchmarkMinSpan/MinSpan_diff_10000-4     	     595	   2397532 ns/op	 1087691 B/op	    8788 allocs/op
BenchmarkMinSpan/MinSpan_diff_53999-4     	     542	   2109992 ns/op	 1087872 B/op	    8789 allocs/op
BenchmarkMaxSpan/MaxSpan_diff_2-4         	      46	  23037379 ns/op	   10817 B/op	      86 allocs/op
BenchmarkMaxSpan/MaxSpan_diff_10-4        	     100	  13874113 ns/op	   12873 B/op	     112 allocs/op
BenchmarkMaxSpan/MaxSpan_diff_100-4       	      50	  22560667 ns/op	   29639 B/op	     209 allocs/op
BenchmarkMaxSpan/MaxSpan_diff_1000-4      	      51	  22326449 ns/op	   36870 B/op	     251 allocs/op
BenchmarkMaxSpan/MaxSpan_diff_10000-4     	      63	  21876837 ns/op	   46030 B/op	     292 allocs/op
BenchmarkMaxSpan/MaxSpan_diff_53999-4     	      62	  21765868 ns/op	   54974 B/op	     350 allocs/op
BenchmarkDetectSpan/Detect_MaxSpan_diff_2-4         	  951639	      1456 ns/op	     544 B/op	      10 allocs/op
BenchmarkDetectSpan/Detect_MaxSpan_diff_10-4        	  985369	      1299 ns/op	     544 B/op	      10 allocs/op
BenchmarkDetectSpan/Detect_MaxSpan_diff_100-4       	 1000000	      1335 ns/op	     544 B/op	      10 allocs/op
BenchmarkDetectSpan/Detect_MaxSpan_diff_1000-4      	 1000000	      1189 ns/op	     544 B/op	      10 allocs/op
BenchmarkDetectSpan/Detect_MaxSpan_diff_10000-4     	 1000000	      1192 ns/op	     544 B/op	      10 allocs/op
BenchmarkDetectSpan/Detect_MaxSpan_diff_53999-4     	 1000000	      1278 ns/op	     544 B/op	      10 allocs/op
BenchmarkDetectSpan/Detect_MinSpan_diff_2-4         	 1000000	      1103 ns/op	     544 B/op	      10 allocs/op
BenchmarkDetectSpan/Detect_MinSpan_diff_10-4        	 1000000	      1078 ns/op	     544 B/op	      10 allocs/op
BenchmarkDetectSpan/Detect_MinSpan_diff_100-4       	 1000000	      1070 ns/op	     544 B/op	      10 allocs/op
BenchmarkDetectSpan/Detect_MinSpan_diff_1000-4      	 1000000	      1079 ns/op	     544 B/op	      10 allocs/op
BenchmarkDetectSpan/Detect_MinSpan_diff_10000-4     	 1000000	      1112 ns/op	     544 B/op	      10 allocs/op
BenchmarkDetectSpan/Detect_MinSpan_diff_53999-4     	 1010179	      1141 ns/op	     544 B/op	      10 allocs/op
**BenchmarkCheckAttestations-4                        	       1	2342823657 ns/op	 2460560 B/op	   22809 allocs/op**

Benchmark After:
BenchmarkMinSpan/MinSpan_diff_2-4         	   10000	   1336615 ns/op	  619177 B/op	    5138 allocs/op
BenchmarkMinSpan/MinSpan_diff_10-4        	   10000	   1363448 ns/op	  619154 B/op	    5138 allocs/op
BenchmarkMinSpan/MinSpan_diff_100-4       	   10000	   1362371 ns/op	  619192 B/op	    5138 allocs/op
BenchmarkMinSpan/MinSpan_diff_1000-4      	   10000	   1368395 ns/op	  619186 B/op	    5138 allocs/op
BenchmarkMinSpan/MinSpan_diff_10000-4     	   10000	   1325822 ns/op	  619162 B/op	    5138 allocs/op
BenchmarkMinSpan/MinSpan_diff_53999-4     	   10000	   1325313 ns/op	  619164 B/op	    5138 allocs/op
BenchmarkMaxSpan/MaxSpan_diff_2-4         	      38	  34699969 ns/op	   11446 B/op	      93 allocs/op
BenchmarkMaxSpan/MaxSpan_diff_10-4        	     100	  23481944 ns/op	   13650 B/op	     118 allocs/op
BenchmarkMaxSpan/MaxSpan_diff_100-4       	      61	  32432117 ns/op	   27191 B/op	     208 allocs/op
BenchmarkMaxSpan/MaxSpan_diff_1000-4      	      42	  33308219 ns/op	   35908 B/op	     254 allocs/op
BenchmarkMaxSpan/MaxSpan_diff_10000-4     	      42	  34042443 ns/op	   44323 B/op	     291 allocs/op
BenchmarkMaxSpan/MaxSpan_diff_53999-4     	      42	  36160023 ns/op	   52700 B/op	     328 allocs/op
BenchmarkDetectSpan/Detect_MaxSpan_diff_2-4         	 1089136	      1227 ns/op	     544 B/op	      10 allocs/op
BenchmarkDetectSpan/Detect_MaxSpan_diff_10-4        	  979192	      1146 ns/op	     544 B/op	      10 allocs/op
BenchmarkDetectSpan/Detect_MaxSpan_diff_100-4       	 1000000	      1178 ns/op	     544 B/op	      10 allocs/op
BenchmarkDetectSpan/Detect_MaxSpan_diff_1000-4      	 1000000	      1212 ns/op	     544 B/op	      10 allocs/op
BenchmarkDetectSpan/Detect_MaxSpan_diff_10000-4     	 1000000	      1137 ns/op	     544 B/op	      10 allocs/op
BenchmarkDetectSpan/Detect_MaxSpan_diff_53999-4     	  953994	      1135 ns/op	     544 B/op	      10 allocs/op
BenchmarkDetectSpan/Detect_MinSpan_diff_2-4         	  980054	      1148 ns/op	     544 B/op	      10 allocs/op
BenchmarkDetectSpan/Detect_MinSpan_diff_10-4        	 1000000	      1252 ns/op	     544 B/op	      10 allocs/op
BenchmarkDetectSpan/Detect_MinSpan_diff_100-4       	 1049521	      1140 ns/op	     544 B/op	      10 allocs/op
BenchmarkDetectSpan/Detect_MinSpan_diff_1000-4      	 1000000	      1233 ns/op	     544 B/op	      10 allocs/op
BenchmarkDetectSpan/Detect_MinSpan_diff_10000-4     	 1000000	      1217 ns/op	     544 B/op	      10 allocs/op
BenchmarkDetectSpan/Detect_MinSpan_diff_53999-4     	 1000000	      1186 ns/op	     544 B/op	      10 allocs/op
**BenchmarkCheckAttestations-4                        	      12	  97340465 ns/op	  848416 B/op	   19559 allocs/op**


Benchmark with update(no batch):
**BenchmarkCheckAttestations-4                        	       1	4854229186 ns/op	 3765640 B/op	   26337 allocs/op**